+++ src/modules/farming/dto/youves-stakes/index.ts
@@ -1,0 +1,2 @@
+export * from './youves-stake.dto';
+export * from './youves-stakes-response.dto';
