+++ src/modules/farming/dto/youves-stakes/youves-stake.dto.ts
