+++ src/modules/farming/dto/youves-stakes/youves-stakes-response.dto.ts
@@ -2,7 +2,7 @@ import { Typed } from '@shared/decorators';
 
 import { YouvesStakeDto } from './youves-stake.dto';
 
-export class YouvesUsersInfoDto {
+export class YouvesStakesResponseDto {
   @Typed({ type: YouvesStakeDto, isArray: true })
   stakes: YouvesStakeDto[];
 }
