--- src/modules/farming/dto/youves-users-info/index.ts
@@ -1,2 +1,0 @@
-export * from './youves-stake.dto';
-export * from './youves-users-info.dto';
