+++ src/modules/farming/models/index.ts
@@ -3,4 +3,4 @@ export * from './farming-list';
 export * from './users-info';
 export * from './user-token-balance.model';
 export * from './youves-farming-item';
-export * from './youves-users-info';
+export * from './youves-stakes';
