+++ src/modules/farming/models/youves-stakes/index.ts
@@ -1,0 +1,2 @@
+export * from './youves-stake.model';
+export * from './youves-stakes-response.model';
