+++ src/modules/farming/models/youves-stakes/youves-stakes-response.model.ts
@@ -1,15 +1,15 @@
-import { YouvesUsersInfoDto } from '../../dto';
+import { YouvesStakesResponseDto } from '../../dto';
 import { YouvesStakeModel } from './youves-stake.model';
 
-export class YouvesUsersInfoModel extends YouvesUsersInfoDto {
+export class YouvesStakesResponseModel extends YouvesStakesResponseDto {
   stakes: YouvesStakeModel[];
 
-  constructor(dto: YouvesUsersInfoDto) {
+  constructor(dto: YouvesStakesResponseDto) {
     super();
 
     for (const key in dto) {
       //@ts-ignore
-      this[key] = dto[key as keyof YouvesUsersInfoDto];
+      this[key] = dto[key as keyof YouvesStakesResponseDto];
     }
 
     this.stakes = dto.stakes.map(item => new YouvesStakeModel(item));
