--- src/modules/farming/models/youves-users-info/index.ts
@@ -1,2 +1,0 @@
-export * from './youves-stake.model';
-export * from './youves-users-info.model';
